Fix #25 handleDot Fix
