+++ src/components/carrousel/utils.ts
@@ -4,31 +4,10 @@ import { FacilityInfo } from './type';
 const dotMarking = (initialNum: number, arr: FacilityInfo[]) => {
   const dotNum = Math.abs(initialNum) / 900;
   for (let idx = 1; idx <= arr.length; idx++) {
-    const elem = window.document.getElementById(`dot_${idx}`);
+    const elem = document.getElementById(`dot_${idx}`);
     if (idx === dotNum && elem) {
       elem.style.backgroundColor = `${Color.GRAY}`;
     } else if (idx !== dotNum && elem) {
-      elem.style.backgroundColor = `${Color.LIGHTGRAY}`;
-    }
-  }
-};
-
-const handleDot = (
-  idx: number,
-  initialNum: number,
-  array: FacilityInfo[],
-  cardListRef: React.MutableRefObject<HTMLDivElement | null>,
-) => {
-  for (let i = 1; i <= array.length; i++) {
-    const elem = document.getElementById(`dot_${i}`);
-    if (i === idx && elem) {
-      elem.style.backgroundColor = `${Color.GRAY}`;
-      initialNum = i * -900;
-      if (cardListRef.current) {
-        cardListRef.current.style.transform = `translateX(${initialNum}px)`;
-        cardListRef.current.style.transitionDuration = '500ms';
-      }
-    } else if (i !== idx && elem) {
       elem.style.backgroundColor = `${Color.LIGHTGRAY}`;
     }
   }
@@ -46,4 +25,4 @@ const addDataAtBothEnds = (array: FacilityInfo[]) => {
   return copyArray;
 };
 
-export { dotMarking, handleDot, addDataAtBothEnds };
+export { dotMarking, addDataAtBothEnds };
